refactor: split place service boundary
