+++ backend/tests/test_page_service.py
@@ -1,18 +1,6 @@
-from types import SimpleNamespace
+def test_page_service_imports():
+    from app.services import page_service
 
-import pytest
-from fastapi import HTTPException, status
-
-from app.services import page_service
-
-
-def test_read_place_service_maps_missing_place_to_404(monkeypatch):
-    def failing_read_place(*_args, **_kwargs):
-        raise ValueError("장소를 찾을 수 없어요.")
-
-    monkeypatch.setattr(page_service, "read_place_entry", failing_read_place)
-
-    with pytest.raises(HTTPException) as caught:
-        page_service.read_place_service(SimpleNamespace(), "missing-place")
-
-    assert caught.value.status_code == status.HTTP_404_NOT_FOUND
+    assert callable(page_service.read_bootstrap_service)
+    assert callable(page_service.read_courses_service)
+    assert callable(page_service.read_reviews_service)
